Further abstraction of GeneralConfig, fixed test config setting
